+++ packages/eas-json/src/schema.ts
@@ -11,11 +11,9 @@ export const EasJsonSchema = Joi.object({
     requireCommit: Joi.boolean(),
     appVersionSource: Joi.string().valid(...Object.values(AppVersionSource)),
     promptToConfigurePushNotifications: Joi.boolean(),
-  }),
-  build: Joi.object().pattern(Joi.string(), BuildProfileSchema),
-  submit: Joi.object().pattern(Joi.string(), UnresolvedSubmitProfileSchema),
-  update: Joi.object({
     assetHostOverride: Joi.string(),
     manifestHostOverride: Joi.string(),
   }),
+  build: Joi.object().pattern(Joi.string(), BuildProfileSchema),
+  submit: Joi.object().pattern(Joi.string(), UnresolvedSubmitProfileSchema),
 });